Remove `call` player action, format code, update readme
Call player action is not needed, because it's implicit in `bid` action.
Bid action represents three actions: call, raise and all in. Server
figures out which is which based on the amount of chips being bid.

Added `mix format`.

Readme updated to include two sample run commands for two players.